Made median adaptive thresholding, fixed normalization and some small bugs

+++ fouriertransform.h
@@ -16,10 +16,10 @@ public:
     static FrequencyAudioData dft(const TimeAudioData &data, size_t start, size_t end);
 
     //Cooley-Tukey fast discrete fourier transform with O(NlogN) of data in range [start, end)
-    static FrequencyAudioData cooleyTukeyFFT(const TimeAudioData &data, size_t start, size_t end, unsigned int fftSizeFactor = 16);
+    static FrequencyAudioData cooleyTukeyFFT(const TimeAudioData &data, size_t start, size_t end,int fftSizeFactor = 16);
 
 private:
-    static size_t reverseBitOrder(size_t a);
+    static quint64 reverseBitOrder(quint64 a);
     static void ctfft(std::vector<Complex> &x);
 };
